+++ src/Graph.ts
@@ -12,9 +12,10 @@ import relativeId from './utils/relativeId';
 import error from './utils/error';
 import { isAbsolute, isRelative, normalize, relative, resolve } from './utils/path';
 import {
+	CachedChunk,
+	CachedChunkSet,
 	InputOptions,
 	IsExternalHook,
-	MissingExportHook,
 	Plugin,
 	ResolveIdHook,
 	RollupWarning,
@@ -69,7 +70,12 @@ export default class Graph {
 	isPureExternalModule: (id: string) => boolean;
 	legacy: boolean;
 	load: (id: string) => Promise<SourceDescription | string | void>;
-	handleMissingExport: MissingExportHook;
+	handleMissingExport: (
+		exportName: string,
+		importingModule: Module,
+		importedModule: string,
+		importerStart?: number
+	) => void;
 	moduleById: Map<string, Module | ExternalModule>;
 	modules: Module[];
 	onwarn: WarningHandler;
@@ -86,9 +92,18 @@ export default class Graph {
 	constructor(options: InputOptions) {
 		this.cachedModules = new Map();
 		if (options.cache) {
-			options.cache.modules.forEach(module => {
-				this.cachedModules.set(module.id, module);
-			});
+			if ((<CachedChunk>options.cache).modules) {
+				(<CachedChunk>options.cache).modules.forEach(module => {
+					this.cachedModules.set(module.id, module);
+				})
+			} else {
+				const chunks = (<CachedChunkSet>options.cache).chunks;
+				for (const chunkName in chunks) {
+					chunks[chunkName].modules.forEach(module => {
+						this.cachedModules.set(module.id, module);
+					});
+				}
+			}
 		}
 		delete options.cache; // TODO not deleting it here causes a memory leak; needs further investigation
 
@@ -144,6 +159,16 @@ export default class Graph {
 			this.plugins
 				.map(plugin => plugin.missingExport)
 				.filter(Boolean)
+				.map(missingExport => {
+					return (
+						exportName: string,
+						importingModule: Module,
+						importedModule: string,
+						importerStart?: number
+					) => {
+						return missingExport(importingModule.id, exportName, importedModule, importerStart);
+					};
+				})
 				.concat(handleMissingExport)
 		);