+++ src/Module.ts
@@ -633,9 +633,9 @@ export default class Module {
 
 			if (!declaration) {
 				this.graph.handleMissingExport(
-					this,
 					importDeclaration.name,
-					otherModule,
+					this,
+					otherModule.id,
 					importDeclaration.specifier.start
 				);
 			}
@@ -665,9 +665,9 @@ export default class Module {
 
 			if (!declaration) {
 				this.graph.handleMissingExport(
-					this,
 					reexportDeclaration.localName,
-					reexportDeclaration.module,
+					this,
+					reexportDeclaration.module.id,
 					reexportDeclaration.start
 				);
 			}